Deleted the main menu xml and java file as not needed.
This branch only contains the profile xml and java file.
